requirements.txt
web3==6.11.1
etherscan-python==2.1.0
requests==2.31.0